Mark up the title more strongly.
It would be nice if the title were to be formatted a little differently than the content to emphasize that it is the title.  This proposes that it be marked up with `<strong>`, which is the same thing that [Mastodon does for headers](https://docs.joinmastodon.org/spec/activitypub/#sanitization).